moved china example
