+++ MapboxAndroidDemo/src/main/AndroidManifest.xml
@@ -674,7 +674,7 @@
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity" />
         </activity>
         <activity
-            android:name=".examples.china.SimpleChinaMapViewActivity"
+            android:name=".examples.SimpleChinaMapViewActivity"
             android:label="@string/activity_china_simple_china_mapview_title"
             android:screenOrientation="portrait">
             <meta-data
